fix(types): moved types into the right location and imported into fromFlux
