+++ giraffe/src/types/index.ts
@@ -614,3 +614,17 @@ export interface BandIndexMap {
   minIndices: number[]
   maxIndices: number[]
 }
+
+export interface Tag {
+  [tagName: string]: string[]
+}
+
+export interface SchemaValues {
+  fields: string[]
+  tags: Tag
+  type?: string
+}
+
+export interface Schema {
+  [measurement: string]: SchemaValues
+}
